refactor: [ANDROAPP-4844] Move ImageDetailBottomDialog.kt to commons
