+++ app/src/main/java/org/dhis2/usescases/searchTrackEntity/mapView/SearchTEMap.kt
@@ -16,6 +16,7 @@ import org.dhis2.Bindings.dp
 import org.dhis2.animations.CarouselViewAnimations
 import org.dhis2.commons.bindings.clipWithRoundedCorners
 import org.dhis2.commons.data.RelationshipOwnerType
+import org.dhis2.commons.dialogs.imagedetail.ImageDetailBottomDialog
 import org.dhis2.commons.locationprovider.LocationSettingLauncher
 import org.dhis2.commons.resources.ColorUtils
 import org.dhis2.databinding.FragmentSearchMapBinding
@@ -32,7 +33,6 @@ import org.dhis2.usescases.searchTrackEntity.SearchTEContractsModule
 import org.dhis2.usescases.searchTrackEntity.SearchTEIViewModel
 import org.dhis2.usescases.searchTrackEntity.SearchTeiViewModelFactory
 import org.dhis2.utils.NetworkUtils
-import org.dhis2.utils.customviews.ImageDetailBottomDialog
 import org.dhis2.utils.isPortrait
 
 const val ARG_FROM_RELATIONSHIP = "ARG_FROM_RELATIONSHIP"
