+++ commons/src/main/res/layout/detail_image_bottom_dialog.xml
